+++ tests/phpstan-bootstrap.php
@@ -7,7 +7,7 @@ if (!function_exists('__')) {
 }
 
 if (!defined('PLUGIN_TORAH_VERSION')) {
-   define('PLUGIN_TORAH_VERSION', '0.4.5');
+   define('PLUGIN_TORAH_VERSION', '0.4.6');
 }
 
 class Config